CorelDraw Autophoto
This Visual Basic macro script is for automatically adding the photo for Library ID cards after the Print merge has been applied.

Create the Database for Print merge in CSV format.
Put all photos in one folder, and make sure the name of the photograph exactly matches the registration number of the user.

Add fileds on the card, with print merge commnad, in CorelDraw

Print merge to a new document.

after printmerge is done, in the new file, (uscually its Printmerge 1) plress Alt F11, to open Macro manager tab. create a new macro in the newly print merged file, paste this code in the file and close the macro window.

rightclick on the newly created macro (Autophoto) in the macro manager pannel.

it will prompt for the object id.. give the object id number of the photo frace(power clipped) where you want the photo to be inserted.

then it will prompt for the folder where all the jpg images are placed. browse to that location.

finish the wizard.

it will take few moments, and after that you will have photos on the respective cards.