Create photo_rename.py
This is a Python script that will read old file names and new file names form the csv file. and then rename the jpg images as per the new filename defined in the csv.

make changes to 'folder_path' and 'csv_file_path' variable to match the folder and file location as per your computer before running the file.